Add a few hacks for NOAA-WOD-PoC
The changes in this commit should never make it to a production instance
of this plugin. They're just hacky things I had to do to get this
proof of concept (PoC) up and running, in part because there are things
about the S3 protocol I don't yet understand.

In particular, I had to add support for cases when no access/secret keys
are provided to the plugin, as they are not needed in the case of the
public bucket I'm interfacing with (and providing them causes failure,
I think because of the way the signatures are provided to the bucket).

I also messed with the way the object names are parsed because I noticed
that if I wanted an object `my/object` from the bucket, the old code parsed
the object as `my`, dropping everything after the `/` delimiter in the object
name.